more or less final version of displaying FieldAccesses in MethodGraphLayout
Commented assertions an debug messages in EdgeDrawer for debugging new
methods.
Overwrites the getSize function in class JoanaVertex for setting the
size of the FieldAccessGraph.
FieldAccessGraph got temporary methods for geting in- and out-Vertices.
MethodGraph Overrides methods getVertexSet, getEdgeSet, got methods for
adding and removing vertices and edges, and for restoring the graph to
its beginning vertex- and edgesets