Added new interface and class, adjusted classes for layouting FieldAccesses
Added Interface DirectedSupplementEdgePath and class
DefaultDirectedSupplementEdgePath for representing a supplementPath also
in context of a normal graph, not just in sugiyama.
Added new function drawEdgesNew to SugiyamaLayoutAlgorithm.
Modified one constructor of SugiyamaGraph for processing graphs that
just need their edges to be drawn new. (fixed coordinates, so he sets
layers...)