fixes #61
Current list is now backuped when opening the dialog, if it is not
accepted, the backup is restored. (fixes changes in priority too)
Color of the color picker is now only saved when vertices get
colored(accepting the dialog).
Also bug fix, removing the last list element caused it to still be
displayed, in the list.