Fixed visual issue #106
Dummies of inter-procedural edges have been drawn with an offset of
there actual position.
This was due to the fact that the content in the shape, namely the empty
label used more space than the visual bounds of the vertex.
Problem was resolved by removing the label when the text is empty.

A better fix would be clipping the label when it's bigger than the vertex.
Enlarging the vertex should not be considered as a solution,
since that could conflict with the layout, which has no way of checking
the size of the label.