Changed the way borders of vertices are applied. fixes #59
The fx-effect drop shadow is no longer used.
Instead the border is applied by backgrounds with insets (which
seems to be the common way to apply borders inside a Pane).
Additionally multiple borders can be applied now.
Therefore the user can have a vertex in multiple groups and see all
borders. And the user can see in which groups a node is while it is
selected.